+++ demo/src/config/navConfig.js
@@ -1,11 +1,14 @@
 /* globals APP_VENDOR */
 import { Video } from '../../schemas';
 
-const mainNav = [
+const navConfig = [
   {
     navType: 'item',
     navId: 'Dashboard',
     to: '/dashboard',
+    redirect: {
+      to: '/dashboard/news',
+    },
   },
   {
     navType: 'dropdown',
@@ -150,11 +153,11 @@ const mainNav = [
 ];
 
 if (Video.schema().hasMixin('triniti:ovp.medialive:mixin:has-channel')) {
-  mainNav[1].dpLinks.push({
+  navConfig[1].dpLinks.push({
     to: '/ovp/livestreams',
     linkTitle: 'Livestreams',
     permission: 'cms-view-videos',
   });
 }
 
-export default mainNav;
+export default navConfig;