+++ demo/src/plugins/dashboard/routes.js
@@ -4,13 +4,19 @@ const r = (id) => `${PLUGIN_PREFIX}${id}`;
 
 const routes = {
   [r('dashboard')]: {
-    path: '/dashboard',
+    path: '/dashboard/:tab(news|active-edits)/',
     component: import('./screens/dashboard'),
+  },
+  [r('dashboardRedirect')]: {
+    path: '/dashboard',
+    redirect: {
+      to: '/dashboard/news',
+    },
   },
   [r('indexRedirect')]: {
     path: '/',
     redirect: {
-      to: '/dashboard',
+      to: '/dashboard/news',
     },
   },
 };