+++ src/plugins/blocksmith/components/pinterest-pin-block-modal/getPinterestPinUrl.js
@@ -1,0 +1,18 @@
+const PINTEREST_PIN_URL_REGEX = /http?s:\/\/(www\.)?pinterest\.com\/pin\/\w+/g;
+const PINTEREST_PIN_ID_REGEX = /^[0-9]*$/g;
+
+let isValidPinUrl = false;
+let isValidPinId = false;
+
+const getPinterestPinUrl = (pin) => {
+  isValidPinUrl = PINTEREST_PIN_URL_REGEX.test(pin);
+  isValidPinId = PINTEREST_PIN_ID_REGEX.test(pin);
+
+  if (isValidPinUrl) {
+    return pin.match(PINTEREST_PIN_URL_REGEX)[0];
+  }
+
+  return isValidPinId ? `https://pinterest.com/pin/${pin}` : '';
+};
+
+export default getPinterestPinUrl;