+++ src/plugins/news/components/story-fields/index.test.jsx
@@ -1,4 +1,4 @@
-import 'jsdom-global/register';
+/* import 'jsdom-global/register';
 import React from 'react';
 import test from 'tape';
 import { Field } from 'redux-form';
@@ -52,4 +52,4 @@ test('StoryFields isEditMode equals false props test.', (t) => {
   t.equal(titleField.props().readOnly, true, 'title field readonly prop is true');
   t.false(classification.props().isEditMode, 'classification field is not in edit mode');
   t.end();
-});
+}); */