+++ src/plugins/raven/index.js
@@ -12,7 +12,7 @@ import Raven from './services/Raven';
 
 export default class RavenPlugin extends Plugin {
   constructor() {
-    super('triniti', 'raven', '0.2.16');
+    super('triniti', 'raven', '0.3.0');
   }
 
   configure(app, bottle) {