+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # CHANGELOG
 
 
+## v0.3.0
+* Make `UserPickerField` and use for `author_ref`
+* Issue: Add Imgur Post Block
+* issue: Add Spotify Embed Block
+* include status in `MediaLiveChannelCards` Card headers.
+* Added tabs for dashboard (News and Active)
+* MediaLive tweaks: increase delay waiting for event, add toast message, and add "no videos found" message
+* Removed asides from code, divider, heading, iframe, page-breakquote and text blocks.
+* Added Youtube Playlist Block
+* Added TikTok embed block
+* Use `@triniti/schemas` v1.1.5
+* Added Pinterest pin embed block
+* MediaLive UI Enhancements
+* Add Slotting On Teasers Screen
+* Add checkbox to enable or disable SmartNews.
+* Remove livestream screen breadcrumbs
+
+
 ## v0.2.16
 * Fix mediaLive reducer bug where endpoints can be duplicated.